+++ docs/releasenotes/changelog.md
@@ -13,6 +13,7 @@
 
 - Fix directories from the ``.gitignore`` file not ignored ([issue #1503](https://github.com/MarketSquare/robotframework-robocop/issues/1503))
 - Fix ``migrate`` command migrating formatters with ``enabled=False`` from the old transform to select option ([issue #1492](https://github.com/MarketSquare/robotframework-robocop/issues/1492))
+- Fix ``migrate`` command not splitting multiline configurations ([issue #1491](https://github.com/MarketSquare/robotframework-robocop/issues/1491))
 - Fix multiline inline IF splitting. To avoid issues when formatting such code, **all inline IFs are now flattened to a single line** ([issue #1506](https://github.com/MarketSquare/robotframework-robocop/issues/1506)):
 
 ```robotframework